+++ src/services/root-reducer.ts
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import { BurgerIngredientsSlice } from './burgeringredients/slice';
+
+export const rootReducer = combineReducers({
+    burgerIngredients: BurgerIngredientsSlice.reducer
+});
