+++ src/services/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { rootReducer } from './reducer';
+import { rootReducer } from './root-reducer';
 
 export const store = configureStore({
     reducer: rootReducer,
